First Draft of @fromQuery; Change Handling of this in Server
Perviously, I was trying to bind this in the @get decorator function.
Now, when a route handling method is defined, the constructor name is
stored with the route mapping and the `this` reference in the RefStore
is retrieve with the name and applied to the called method in the server.